requirements.txt
alpaca-trade-api
pandas
python-dotenv
numpy
pandas_market_calendars
